Fix issue https://github.com/GluuFederation/oxAuth/issues/64
